+++ CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## 22.4.1
 
-### Additions and Improvements
+### Breaking Changes
+- To use clique consensus, require --Xclique-enabled option [] (https://github.com/hyperledger/besu/pull/)
+
+- ### Additions and Improvements
 - GraphQL - allow null log topics in queries which match any topic [#3662](https://github.com/hyperledger/besu/pull/3662)
 - multi-arch docker builds for amd64 and arm64 [#2954](https://github.com/hyperledger/besu/pull/2954)
 - Filter Netty native lib errors likewise the pure Java implementation [#3807](https://github.com/hyperledger/besu/pull/3807)